float

+++ dashboard/models.py
@@ -9,6 +9,6 @@ class Holiday(models.Model):
 class Info(models.Model):
     name = models.CharField(max_length=20, default='main')    
     geographic_coords = models.CharField(max_length=75)
-    radius = models.IntegerField()
+    radius = models.FloatField()
     time_in = models.TimeField(null=True, blank=True)
     time_out = models.TimeField(null=True, blank=True)
